handle most cases in parsing

+++ libft/ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: aheddak <aheddak@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/12 20:50:07 by aheddak           #+#    #+#             */
-/*   Updated: 2022/09/23 22:22:16 by aheddak          ###   ########.fr       */
+/*   Updated: 2022/10/12 05:42:29 by aheddak          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,20 +14,20 @@
 
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
-	size_t		i;
-	char		*dest;
+	int	i;
+	char			*dest;
 
 	i = 0;
 	if (!s)
 		return (NULL);
-	if (!len || start >= ft_strlen(s))
+	if (!len || start >= (unsigned int)ft_strlen(s))
 		return (ft_strdup(""));
-	if (len > ft_strlen(s))
+	if (len > (size_t)ft_strlen(s))
 		len = ft_strlen(s);
 	dest = (char *) malloc (sizeof(char) * (len + 1));
 	if (dest == 0)
 		return (NULL);
-	while (i < len)
+	while ((size_t)i < len)
 	{
 		dest[i] = s[start + i];
 		i++;
